* EOControl/EOEditingContext.[h|m]           switch setLevelsOfUndo to NSUInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@36717 72102866-910b-0410-8b05-ffd578937521

+++ EOControl/EOEditingContext.h
@@ -121,7 +121,7 @@
 - (void)insertObject: (id)object
         withGlobalID: (EOGlobalID *)gid;
 
--(void)setLevelsOfUndo: (int)levels;
+-(void)setLevelsOfUndo: (NSUInteger)levels;
 
 - (void)deleteObject: (id)object;
 - (void)lockObject: (id)object;
